+++ packages/jsonapi/src/ResourceConfig/Builder/AbstractResourceConfigBuilder.php
@@ -22,6 +22,7 @@ use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Wrapping\PropertyBehavior\ConstructorBehaviorInterface;
 use EDT\Wrapping\PropertyBehavior\PropertySetBehaviorInterface;
 use EDT\Wrapping\PropertyBehavior\PropertyUpdatabilityInterface;
+use function array_key_exists;
 
 /**
  * @template TCondition of PathsBasedInterface
@@ -35,17 +36,17 @@ abstract class AbstractResourceConfigBuilder implements ResourceConfigBuilderInt
     /**
      * @var array<non-empty-string, AttributeConfigBuilder<TCondition, TEntity>>
      */
-    protected array $attributes = [];
+    private array $attributes = [];
 
     /**
      * @var array<non-empty-string, ToOneRelationshipConfigBuilder<TCondition, TSorting, TEntity, object>>
      */
-    protected array $toOneRelationships = [];
+    private array $toOneRelationships = [];
 
     /**
      * @var array<non-empty-string, ToManyRelationshipConfigBuilder<TCondition, TSorting, TEntity, object>>
      */
-    protected array $toManyRelationships = [];
+    private array $toManyRelationships = [];
 
     /**
      * @var list<ConstructorBehaviorInterface>
@@ -89,6 +90,14 @@ abstract class AbstractResourceConfigBuilder implements ResourceConfigBuilderInt
         return $this->attributes[$propertyName] ?? null;
     }
 
+    /**
+     * @param non-empty-string $propertyName
+     */
+    protected function hasAttributeConfigBuilder(string $propertyName): bool
+    {
+        return array_key_exists($propertyName, $this->attributes);
+    }
+
     public function setAttributeConfigBuilder(string $propertyName, AttributeConfigBuilder $builder): void
     {
         $this->attributes[$propertyName] = $builder;
@@ -99,6 +108,14 @@ abstract class AbstractResourceConfigBuilder implements ResourceConfigBuilderInt
         return $this->toOneRelationships[$propertyName] ?? null;
     }
 
+    /**
+     * @param non-empty-string $propertyName
+     */
+    protected function hasToOneRelationshipConfigBuilder(string $propertyName): bool
+    {
+        return array_key_exists($propertyName, $this->toOneRelationships);
+    }
+
     public function setToOneRelationshipConfigBuilder(string $propertyName, ToOneRelationshipConfigBuilder $builder): void
     {
         $this->toOneRelationships[$propertyName] = $builder;
@@ -107,6 +124,14 @@ abstract class AbstractResourceConfigBuilder implements ResourceConfigBuilderInt
     public function getToManyRelationshipConfigBuilder(string $propertyName): ?ToManyRelationshipConfigBuilderInterface
     {
         return $this->toManyRelationships[$propertyName] ?? null;
+    }
+
+    /**
+     * @param non-empty-string $propertyName
+     */
+    protected function hasToManyRelationshipConfigBuilder(string $propertyName): bool
+    {
+        return array_key_exists($propertyName, $this->toManyRelationships);
     }
 
     public function setToManyRelationshipConfigBuilder(string $propertyName, ToManyRelationshipConfigBuilder $builder): void